mrc-layer-setup(small-funcs-move,mechanical_provable): Extract PPLayerRange / compute_model_num_layers / resolve_pp_layer_range / assert_pp_mtp_compat to layer_setup.py (cut+paste)
